requirements.txt
APScheduler==3.9.1
arrow==1.2.2
asgiref==3.5.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
blessed==1.19.1
bs4==0.0.1
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.5
django-apscheduler==0.6.2
django-picklefield==3.1
djangorestframework==3.13.1
idna==3.3
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
redis==3.5.3
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
wcwidth==0.2.5

